Enable torch models
