+++ nnlo/models/TorchModels.py
@@ -1,9 +1,6 @@
 import torch
 from torch.autograd import Variable
 import torch.nn as nn
-import torch.nn.parallel
-import torch.backends.cudnn as cudnn
-import torch.distributed as dist
 import torch.optim
 import torch.utils.data.distributed
 import torchvision.transforms as transforms
@@ -12,32 +9,6 @@ import torchvision.models as models
 import torch.nn.functional as F
 import numpy
 import logging
-
-class MNistNet(nn.Module):
-    def __init__(self, **args):
-        super(MNistNet, self).__init__()
-        ks = int(args.get('kernel_size',5))
-        do = float(args.get('dropout',0.5))
-        dense = int(args.get('dense',50))
-        self.conv1 = nn.Conv2d(1, 10, kernel_size=ks)
-        self.conv2 = nn.Conv2d(10, 20, kernel_size=ks)
-        self.conv2_drop = nn.Dropout2d(do)
-        self.fc1 = nn.Linear(320, dense)
-        self.fc2 = nn.Linear(dense, 10)
-
-    def forward(self, x):
-        x = x.permute(0,3,1,2).float()
-        x = F.relu(F.max_pool2d(self.conv1(x), 2))
-        x = F.relu(F.max_pool2d(self.conv2_drop(self.conv2(x)), 2))
-        x = x.view(-1, 320)
-        x = F.relu(self.fc1(x))
-        x = F.dropout(x, training=self.training)
-        x = self.fc2(x)
-        #return F.log_softmax(x, dim=1)
-        #return F.softmax(x)
-        #return F.cross_entropy(x)
-        return x
-
 
 
 ### Build a customized CNN with given hyperparameters
